cleanup crypto pkg
